Move Validator to root namespace
