--- sig/kangaru/validators/validator.rbs
@@ -1,20 +1,0 @@
-module Kangaru
-  module Validators
-    class Validator
-      attr_reader model: untyped
-      attr_reader attribute: Symbol
-      attr_reader params: Hash[Symbol, untyped]
-      attr_reader value: untyped
-
-      def initialize: (
-        model: untyped,
-        attribute: Symbol,
-        params: Hash[Symbol, untyped]
-      ) -> void
-
-      def validate: -> void
-
-      def add_error!: (Symbol) -> void
-    end
-  end
-end
