+++ spec/kangaru/validator_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Kangaru::Validators::Validator do
+RSpec.describe Kangaru::Validator do
   subject(:validator) { described_class.new(**attributes) }
 
   let(:attributes) { { model:, attribute:, params: } }
